Se insertó funcionalidad de eliminación de usuario por id en su respectiva carpeta para separar responsabilidades
