+++ gym-back-end/src/admin/services/user.service.ts
@@ -1,4 +1,4 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from 'src/common/models/entities/user.entity';
 import { IUserResponse } from 'src/common/models/interfaces/user.interface';
@@ -6,6 +6,7 @@ import { Repository } from 'typeorm';
 
 @Injectable()
 export class UserService {
+    private readonly logger = new Logger(UserService.name);
 
     constructor(@InjectRepository(User) private readonly userRepository: Repository<User>) { }
 
@@ -21,6 +22,27 @@ export class UserService {
         } catch (error) {
             console.error("Error al traer el listado de usuarios:", error);
             throw new HttpException({ message: 'Ocurrió un error para encontrar el listado de usuarios' }, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    async deleteUserById(id: number): Promise<{ message: string; affected: number }> {
+        try {
+            const userDeleted = await this.userRepository.delete({ usuario_id: id });
+            if (userDeleted.affected === 0) {
+                this.logger.warn(`Usuario con id ${id} no encontrado`);
+                throw new NotFoundException(`Usuario con id ${id} no encontrado`);
+            }
+            this.logger.log(`Usuario con id ${id} eliminado correctamente!`);
+            return { message: `Usuario con id ${id} eliminado correctamente`, affected: userDeleted.affected };
+        } catch (error) {
+            this.logger.error(`Error al eliminar el usuario con id ${id}`, error.stack);
+            if (error instanceof HttpException) {
+                throw error;
+            }
+            throw new HttpException(
+                `Error inesperado al eliminar el usuario con el ID ${id}`,
+                HttpStatus.INTERNAL_SERVER_ERROR,
+            );
         }
     }
 }
